Make libraries.sh work with pkgname and provides
It will now look for all PSPBUILDs and send the package name to
build_package function instead of the directory. build_package will then
look for a PSPBUILD which either has this package in pkgname or
provides.

This is a bit inefficient, but it will make us able to use provides.