Initial support for building prx files with CMAKE
This is still WIP, because I'm still having some issues with it, but
a prx is being build. I haven't been able to launch the prx in psplink
for one the projects I have tested this on, so I may need some help
debugging this.

Using this is quite simple, just use the ``build_prx(binary)`` in your
CMakeLists.txt. Do make sure to replace binary with your binary target.